firebase/functions/requirements.txt
firebase-functions
firebase-admin
Pillow
onnxruntime>=1.16.0
tensorflow
numpy